requirements.txt
Pillow==10.3.0
py2app==0.28.6
tk
